Fix all base_unittests for m70 base library!
I disabled only a handful of tests. These test cases test features
we don't have/need(like shared memory), a few logging test wants\
to test DCHECK but Starboard DCHECK doesn't provide ability to
switch on and off(so these tests crash).

b/27798606

+++ base/files/file_starboard.cc
@@ -155,7 +155,7 @@ int File::Write(int64_t offset, const char* data, int size) {
     return -1;
   }
 
-  int position = SbFileSeek(file_.get(), kSbFileFromBegin, offset);
+  int64_t position = SbFileSeek(file_.get(), kSbFileFromBegin, offset);
   int result = 0;
   if (position == offset) {
     result = WriteAtCurrentPos(data, size);
